+++ crd-generator/api/src/test/java/io/fabric8/crd/generator/v1/JsonSchemaTest.java
@@ -18,6 +18,7 @@ package io.fabric8.crd.generator.v1;
 import com.fasterxml.jackson.databind.JsonNode;
 import io.fabric8.crd.example.annotated.Annotated;
 import io.fabric8.crd.example.basic.Basic;
+import io.fabric8.crd.example.json.ContainingJson;
 import io.fabric8.crd.example.person.Person;
 import io.fabric8.crd.generator.utils.Types;
 import io.fabric8.kubernetes.api.model.apiextensions.v1.JSONSchemaProps;
@@ -103,5 +104,32 @@ class JsonSchemaTest {
     final List<JsonNode> enumValues = anEnum.getEnum();
     assertEquals(2, enumValues.size());
     enumValues.stream().map(JsonNode::textValue).forEach(s -> assertTrue("oui".equals(s) || "non".equals(s)));
+
+    // check ignored fields
+    assertFalse(spec.containsKey("ignoredFoo"));
+    assertFalse(spec.containsKey("ignoredBar"));
+  }
+
+  @Test
+  void shouldProbuceKubernetesPreserveFields() {
+    TypeDef containingJson = Types.typeDefFrom(ContainingJson.class);
+    JSONSchemaProps schema = JsonSchema.from(containingJson);
+    assertNotNull(schema);
+    Map<String, JSONSchemaProps> properties = schema.getProperties();
+    assertEquals(2, properties.size());
+    final JSONSchemaProps specSchema = properties.get("spec");
+    Map<String, JSONSchemaProps> spec = specSchema.getProperties();
+    assertEquals(2, spec.size());
+
+    // check descriptions are present
+    assertTrue(spec.containsKey("free"));
+    JSONSchemaProps freeField = spec.get("free");
+
+    assertTrue(freeField.getXKubernetesPreserveUnknownFields());
+
+    assertTrue(spec.containsKey("field"));
+    JSONSchemaProps field = spec.get("field");
+
+    assertNull(field.getXKubernetesPreserveUnknownFields());
   }
 }