+++ kubernetes-itests/src/test/java/io/fabric8/kubernetes/APIServiceIT.java
@@ -15,6 +15,7 @@
  */
 package io.fabric8.kubernetes;
 
+import io.fabric8.commons.AssumingK8sVersionAtLeast;
 import io.fabric8.commons.ClusterEntity;
 import io.fabric8.kubernetes.api.model.APIService;
 import io.fabric8.kubernetes.api.model.APIServiceBuilder;
@@ -22,20 +23,25 @@ import io.fabric8.kubernetes.api.model.APIServiceList;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import org.arquillian.cube.kubernetes.impl.requirement.RequiresKubernetes;
 import org.arquillian.cube.requirement.ArquillianConditionalRunner;
+import org.assertj.core.api.InstanceOfAssertFactories;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
+import org.junit.ClassRule;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static junit.framework.TestCase.assertTrue;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
+import static org.assertj.core.api.Assertions.assertThat;
 
 @RunWith(ArquillianConditionalRunner.class)
 @RequiresKubernetes
 public class APIServiceIT {
+
+  @ClassRule
+  public static final AssumingK8sVersionAtLeast assumingK8sVersion =
+    new AssumingK8sVersionAtLeast("1", "16");
+
   @ArquillianResource
   KubernetesClient client;
 
@@ -44,36 +50,42 @@ public class APIServiceIT {
     ClusterEntity.apply(APIServiceIT.class.getResourceAsStream("/apiservice-it.yml"));
   }
 
+  @AfterClass
+  public static void cleanup() {
+    ClusterEntity.remove(APIServiceIT.class.getResourceAsStream("/apiservice-it.yml"));
+  }
+
   @Test
   public void get() {
-    APIService apiService = client.apiServices().withName("v1beta1.get.fabric8.io").get();
-    assertThat(apiService).isNotNull();
+    APIService result = client.apiServices().withName("v1.tests.example.com").get();
+    assertThat(result)
+      .hasFieldOrPropertyWithValue("spec.group", "tests.example.com")
+      .hasFieldOrPropertyWithValue("spec.version", "v1");
   }
 
   @Test
   public void list() {
-    APIServiceList aAPIServiceList = client.apiServices().list();
-    assertNotNull(aAPIServiceList);
-    assertTrue(aAPIServiceList.getItems().size() >= 1);
+    APIServiceList result = client.apiServices().list();
+    assertThat(result).extracting(APIServiceList::getItems)
+      .asInstanceOf(InstanceOfAssertFactories.list(APIService.class))
+      .anyMatch(t -> t.getMetadata().getName().equals("v1.tests.example.com"));
   }
 
   @Test
   public void update() {
-    APIService apiService = client.apiServices().withName("v1beta1.update.fabric8.io").edit(c -> new APIServiceBuilder(c)
-      .editOrNewMetadata().addToAnnotations("foo", "bar").endMetadata()
-      .build());
-
-    assertNotNull(apiService);
-    assertEquals("bar", apiService.getMetadata().getAnnotations().get("foo"));
+    APIService result = client.apiServices().withName("v1.tests.example.com")
+      .edit(c -> new APIServiceBuilder(c)
+        .editOrNewMetadata().addToAnnotations("foo", "bar").endMetadata()
+        .build());
+    assertThat(result)
+      .hasFieldOrPropertyWithValue("metadata.annotations.foo", "bar")
+      .hasFieldOrPropertyWithValue("spec.group", "tests.example.com")
+      .hasFieldOrPropertyWithValue("spec.version", "v1");
   }
 
   @Test
+  @Ignore(value = "https://github.com/fabric8io/kubernetes-client/issues/3668")
   public void delete() {
-    assertTrue(client.apiServices().withName("v1beta1.delete.fabric8.io").delete());
-  }
-
-  @AfterClass
-  public static void cleanup() {
-    ClusterEntity.remove(APIServiceIT.class.getResourceAsStream("/apiservice-it.yml"));
+    assertThat(client.apiServices().withName("v1beta1.delete.fabric8.io").delete()).isTrue();
   }
 }